+++ src/index.js
@@ -1,20 +1,29 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import Switch from 'react-router-dom/Switch';
-import Route from 'react-router-dom/Route';
 import 'core-js/stable';
 import 'regenerator-runtime/runtime';
 
-import MediatedRequests from './routes/MediatedRequests';
+import {
+  Route,
+  Switch,
+  Redirect,
+} from '@folio/stripes/core';
+
+import ConfirmItemArrival from './components/ConfirmItemArrival';
+import MediatedRequestsActions from './components/MediatedRequestsActions';
+import SendItemInTransit from './components/SendItemInTransit';
 import Settings from './settings';
+
+import {
+  CONFIRM_ITEM_ARRIVAL,
+  MEDIATED_REQUESTS_ACTIONS,
+  SEND_ITEM_IN_TRANSIT,
+} from './constants';
 
 class RequestsMediated extends React.Component {
   static propTypes = {
     match: PropTypes.object.isRequired,
     showSettings: PropTypes.bool,
-    stripes: PropTypes.shape({
-      connect: PropTypes.func,
-    })
   };
 
   render() {
@@ -32,10 +41,18 @@ class RequestsMediated extends React.Component {
     return (
       <Switch>
         <Route
-          path={path}
-          exact
-          component={MediatedRequests}
+          component={MediatedRequestsActions}
+          path={`${path}/${MEDIATED_REQUESTS_ACTIONS}`}
         />
+        <Route
+          path={`${path}/${CONFIRM_ITEM_ARRIVAL}`}
+          component={ConfirmItemArrival}
+        />
+        <Route
+          path={`${path}/${SEND_ITEM_IN_TRANSIT}`}
+          component={SendItemInTransit}
+        />
+        <Redirect to={`${path}/${MEDIATED_REQUESTS_ACTIONS}`} />
       </Switch>
     );
   }