+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Change history for ui-requests-mediated
 
 ## 1.1.0 (IN PROGRESS)
-
+* Add landing page actions: Confirm item arrival, Mediated requests actions and Send item in transit. Refs UIREQMED-2.
 
 ## 1.0.0